move context into the another module, make it more general
