+++ jormungandr/src/lib.rs
@@ -31,6 +31,7 @@ use std::time::Duration;
 pub mod blockcfg;
 pub mod blockchain;
 pub mod client;
+pub mod context;
 pub mod diagnostic;
 pub mod fragment;
 pub mod intercom;
@@ -65,7 +66,7 @@ pub struct BootstrappedNode {
     blockchain: Blockchain,
     blockchain_tip: blockchain::Tip,
     block0_hash: HeaderHash,
-    rest_context: Option<rest::ContextLock>,
+    context: Option<context::ContextLock>,
     services: Services,
     initial_peers: Vec<topology::Peer>,
     _logger_guards: Vec<WorkerGuard>,
@@ -80,7 +81,7 @@ const WATCH_CLIENT_TASK_QUEUE_LEN: usize = 32;
 const BOOTSTRAP_RETRY_WAIT: Duration = Duration::from_secs(5);
 
 fn start_services(bootstrapped_node: BootstrappedNode) -> Result<(), start_up::Error> {
-    if let Some(context) = bootstrapped_node.rest_context.as_ref() {
+    if let Some(context) = bootstrapped_node.context.as_ref() {
         block_on(async {
             context
                 .write()
@@ -305,8 +306,8 @@ fn start_services(bootstrapped_node: BootstrappedNode) -> Result<(), start_up::E
         });
     };
 
-    if let Some(rest_context) = bootstrapped_node.rest_context {
-        let full_context = rest::FullContext {
+    if let Some(context) = bootstrapped_node.context {
+        let full_context = context::FullContext {
             stats_counter: simple_metrics_counter,
             network_task: network_msgbox,
             transaction_task: fragment_msgbox,
@@ -318,9 +319,9 @@ fn start_services(bootstrapped_node: BootstrappedNode) -> Result<(), start_up::E
             prometheus: prometheus_metric,
         };
         block_on(async {
-            let mut rest_context = rest_context.write().await;
-            rest_context.set_full(full_context);
-            rest_context.set_node_state(NodeState::Running);
+            let mut context = context.write().await;
+            context.set_full(full_context);
+            context.set_node_state(NodeState::Running);
         })
     };
 
@@ -368,7 +369,7 @@ fn bootstrap(initialized_node: InitializedNode) -> Result<BootstrappedNode, star
         settings,
         block0,
         storage,
-        rest_context,
+        context,
         mut services,
         cancellation_token,
         _logger_guards,
@@ -378,12 +379,12 @@ fn bootstrap(initialized_node: InitializedNode) -> Result<BootstrappedNode, star
         blockchain,
         blockchain_tip,
         block0_hash,
-        rest_context,
+        context,
         settings,
         initial_peers,
     } = services.block_on_task("bootstrap", |info| {
         bootstrap_internal(
-            rest_context,
+            context,
             info.span().clone(),
             block0,
             storage,
@@ -397,7 +398,7 @@ fn bootstrap(initialized_node: InitializedNode) -> Result<BootstrappedNode, star
         blockchain,
         blockchain_tip,
         block0_hash,
-        rest_context,
+        context,
         services,
         initial_peers,
         _logger_guards,
@@ -408,13 +409,13 @@ struct BootstrapData {
     blockchain: Blockchain,
     blockchain_tip: blockchain::Tip,
     block0_hash: HeaderHash,
-    rest_context: Option<rest::ContextLock>,
+    context: Option<context::ContextLock>,
     settings: Settings,
     initial_peers: Vec<topology::Peer>,
 }
 
 async fn bootstrap_internal(
-    rest_context: Option<rest::ContextLock>,
+    context: Option<context::ContextLock>,
     span: Span,
     block0: blockcfg::Block,
     storage: blockchain::Storage,
@@ -423,7 +424,7 @@ async fn bootstrap_internal(
 ) -> Result<BootstrapData, start_up::Error> {
     use futures::future::FutureExt;
 
-    if let Some(context) = rest_context.as_ref() {
+    if let Some(context) = context.as_ref() {
         block_on(async {
             context
                 .write()
@@ -440,7 +441,7 @@ async fn bootstrap_internal(
         start_up::load_blockchain(block0, storage, cache_capacity, settings.rewards_report_all)
             .await?;
 
-    if let Some(context) = &rest_context {
+    if let Some(context) = &context {
         let mut context = context.write().await;
         context.set_blockchain(blockchain.clone());
         context.set_blockchain_tip(blockchain_tip.clone());
@@ -486,7 +487,7 @@ async fn bootstrap_internal(
         }
     };
 
-    if let Some(context) = &rest_context {
+    if let Some(context) = &context {
         let mut context = context.write().await;
         context.remove_bootstrap_stopper();
     };
@@ -495,7 +496,7 @@ async fn bootstrap_internal(
         block0_hash,
         blockchain,
         blockchain_tip,
-        rest_context,
+        context,
         settings,
         initial_peers: network_res
             .map(|res| res.initial_peers)
@@ -507,7 +508,7 @@ pub struct InitializedNode {
     pub settings: Settings,
     pub block0: blockcfg::Block,
     pub storage: blockchain::Storage,
-    pub rest_context: Option<rest::ContextLock>,
+    pub context: Option<context::ContextLock>,
     pub services: Services,
     pub cancellation_token: CancellationToken,
     pub _logger_guards: Vec<WorkerGuard>,
@@ -617,16 +618,18 @@ fn initialize_node() -> Result<InitializedNode, start_up::Error> {
     let cancellation_token = CancellationToken::new();
     init_os_signal_watchers(&mut services, cancellation_token.clone());
 
-    let rest_context = match settings.rest.clone() {
+    let init_context = |diagnostic| {
+        use tokio::sync::RwLock;
+
+        let mut context = context::Context::new();
+        context.set_diagnostic_data(diagnostic);
+        context.set_node_state(NodeState::PreparingStorage);
+        Arc::new(RwLock::new(context))
+    };
+
+    let context = match settings.rest.clone() {
         Some(rest_config) => {
-            use tokio::sync::RwLock;
-
-            let mut context = rest::Context::new();
-            context.set_diagnostic_data(diagnostic);
-            context.set_node_state(NodeState::PreparingStorage);
-            let context = Arc::new(RwLock::new(context));
-
-            let service_context = context.clone();
+            let context = init_context(diagnostic);
 
             let rest_config = rest::Config {
                 listen: rest_config.listen,
@@ -637,6 +640,7 @@ fn initialize_node() -> Result<InitializedNode, start_up::Error> {
             };
 
             let server_handler = rest::start_rest_server(rest_config, context.clone());
+            let service_context = context.clone();
             services.spawn_future("rest", |info| async move {
                 service_context.write().await.set_span(info.span().clone());
                 server_handler.await
@@ -656,7 +660,7 @@ fn initialize_node() -> Result<InitializedNode, start_up::Error> {
 
     // TODO: load network module here too (if needed)
 
-    if let Some(context) = rest_context.as_ref() {
+    if let Some(context) = context.as_ref() {
         block_on(async {
             context
                 .write()
@@ -671,7 +675,7 @@ fn initialize_node() -> Result<InitializedNode, start_up::Error> {
 
         let cancellation_token = CancellationToken::new();
 
-        if let Some(context) = rest_context.as_ref() {
+        if let Some(context) = context.as_ref() {
             let mut context = context.write().await;
             context.set_bootstrap_stopper(cancellation_token.clone());
         }
@@ -683,7 +687,7 @@ fn initialize_node() -> Result<InitializedNode, start_up::Error> {
             _ = cancellation_token.cancelled().fuse() => return Err(start_up::Error::Interrupted),
         };
 
-        if let Some(context) = rest_context.as_ref() {
+        if let Some(context) = context.as_ref() {
             let mut context = context.write().await;
             context.remove_bootstrap_stopper();
         }
@@ -696,7 +700,7 @@ fn initialize_node() -> Result<InitializedNode, start_up::Error> {
         settings,
         block0,
         storage,
-        rest_context,
+        context,
         services,
         cancellation_token,
         _logger_guards,
